% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotated_Library.R
\docType{data}
\name{Annotated_Library}
\alias{Annotated_Library}
\title{Annotated_Library Tables}
\format{list of tibbles
\describe{
\item{lib.neg}{A tibble: 3,978 x 9}
\item{lib.pos}{A tibble: 7,514 x 9}
}}
\source{
https://github.com/jmosl01/lcmsfishdata/blob/master/data-raw/Annotated_Library
}
\usage{
Annotated_Library
}
\description{
Tables generated from the Annotated.library input to search_IHL
}
\examples{
 Annotated_Library
}
\keyword{datasets}
